:recycle: #67 ChatMessage の型定義を厳密に行うように修正
