+++ src/app/chat/_components/ChatContent/ChatContent.stories.tsx
@@ -1,4 +1,4 @@
-import { createInternalApiUrl } from '@/features';
+import { createInternalApiUrl, type ChatMessage } from '@/features';
 import { mockFetchCatMessage } from '@/mocks/api';
 import type { Meta, StoryObj } from '@storybook/react';
 import { http } from 'msw';
@@ -31,7 +31,7 @@ const chatMessages = [
     name: 'User',
     message: 'こんにちはもこちゃん！お話しよう！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -39,14 +39,14 @@ const chatMessages = [
       'こんにちはにゃん🐱もことお話しようにゃん！もこはねこだけど、チュールは苦手だにゃ🐱チキン味のカリカリしか食べないにゃん！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'user',
     name: 'User',
     message:
       'ねこちゃんはチュール好きな子多いけど、もこちゃんは好きじゃないんだね！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -54,13 +54,13 @@ const chatMessages = [
       'そうにゃ🐱もこはウェットフードが苦手だにゃ🐱Userさんの好きな食べ物を教えてにゃー！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'user',
     name: 'User',
     message: 'こんにちはもこちゃん！お話しよう！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -68,14 +68,14 @@ const chatMessages = [
       'こんにちはにゃん🐱もことお話しようにゃん！もこはねこだけど、チュールは苦手だにゃ🐱チキン味のカリカリしか食べないにゃん！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'user',
     name: 'User',
     message:
       'ねこちゃんはチュール好きな子多いけど、もこちゃんは好きじゃないんだね！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -83,13 +83,13 @@ const chatMessages = [
       'そうにゃ🐱もこはウェットフードが苦手だにゃ🐱Userさんの好きな食べ物を教えてにゃー！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'user',
     name: 'User',
     message: 'こんにちはもこちゃん！お話しよう！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -97,14 +97,14 @@ const chatMessages = [
       'こんにちはにゃん🐱もことお話しようにゃん！もこはねこだけど、チュールは苦手だにゃ🐱チキン味のカリカリしか食べないにゃん！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'user',
     name: 'User',
     message:
       'ねこちゃんはチュール好きな子多いけど、もこちゃんは好きじゃないんだね！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -112,13 +112,13 @@ const chatMessages = [
       'そうにゃ🐱もこはウェットフードが苦手だにゃ🐱Userさんの好きな食べ物を教えてにゃー！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'user',
     name: 'User',
     message: 'こんにちはもこちゃん！お話しよう！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -126,14 +126,14 @@ const chatMessages = [
       'こんにちはにゃん🐱もことお話しようにゃん！もこはねこだけど、チュールは苦手だにゃ🐱チキン味のカリカリしか食べないにゃん！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'user',
     name: 'User',
     message:
       'ねこちゃんはチュール好きな子多いけど、もこちゃんは好きじゃないんだね！',
     avatarUrl: 'https://avatars.githubusercontent.com/u/11032365?s=96&v=4',
-  } as const,
+  } as const satisfies ChatMessage,
   {
     role: 'cat',
     name: 'もこちゃん',
@@ -141,7 +141,7 @@ const chatMessages = [
       'そうにゃ🐱もこはウェットフードが苦手だにゃ🐱Userさんの好きな食べ物を教えてにゃー！',
     avatarUrl:
       'https://lgtm-images.lgtmeow.com/2022/03/23/10/9738095a-f426-48e4-be8d-93f933c42917.webp',
-  } as const,
+  } as const satisfies ChatMessage,
 ];
 
 export const Default: Story = {
