+++ src/features/chat.ts
@@ -1,0 +1,8 @@
+export type ChatMessage = {
+  role: 'user' | 'cat';
+  name: string;
+  message: string;
+  avatarUrl: string;
+};
+
+export type ChatMessages = ChatMessage[];
