+++ src/features/index.ts
@@ -5,6 +5,8 @@ export type {
   FetchCatMessage,
 } from './cat';
 
+export type { ChatMessage, ChatMessages } from './chat';
+
 export { extractCatNameById, isCatId, isFetchCatMessageResponse } from './cat';
 
 export { createInternalApiUrl } from './url';
